feat(component): reimplement VocabularySelectField using I18nString & SUI Breadcrumb
