+++ oarepo_vocabularies/ui/theme/assets/semantic-ui/js/oarepo_vocabularies_ui/form/components/index.js
@@ -6,4 +6,4 @@ export * from './PublishButton'
 export * from './ResetButton'
 export * from './VocabularyBreadcrumb'
 export * from './VocabularyBreadcrumbMessage'
-export * from './VocabularySelect'
+export * from './VocabularySelectField'
